+++ CP/src/heuristics/heuristics.hpp
@@ -14,9 +14,10 @@ using namespace Gecode;
 using namespace Search;
 
 /// Custom includes
+#include "solvers/box_wrapper_optim.hpp"
 #include "utils/wrapped_boxes.hpp"
 #include "utils/input_data.hpp"
-#include "solvers/box_wrapper_optim.hpp"
+#include "utils/definitions.hpp"
 
 namespace heuristics {
 	
@@ -24,7 +25,7 @@ namespace heuristics {
 		bool call_solver
 		(
 			Options& ops, const gifts& data,
-			size_t stop_at,
+			uint32 stop_at,
 			bool stop_time, double stop_when,
 			bool verbose,
 			wrapped_boxes& BEST, bool& error
@@ -34,7 +35,7 @@ namespace heuristics {
 	bool heuristic_inc
 	(
 		Options& search_options,
-		size_t stop_at,
+		uint32 stop_at,
 		bool stop_time, double stop_when,
 		bool verbose,
 		const gifts& input_data,
@@ -44,7 +45,7 @@ namespace heuristics {
 	bool heuristic_dec
 	(
 		Options& search_options,
-		size_t stop_at,
+		uint32 stop_at,
 		bool stop_time, double stop_when,
 		bool verbose,
 		const gifts& input_data,
@@ -54,7 +55,7 @@ namespace heuristics {
 	bool heuristic_rand
 	(
 		Options& search_options,
-		size_t rand_times, size_t stop_at,
+		uint32 rand_times, uint32 stop_at,
 		bool stop_time, double stop_when,
 		bool verbose,
 		const gifts& input_data,
@@ -64,7 +65,7 @@ namespace heuristics {
 	bool heuristic_mix
 	(
 		Options& search_options,
-		size_t rand_times, size_t stop_at,
+		uint32 rand_times, uint32 stop_at,
 		bool stop_time, double stop_when,
 		bool verbose,
 		const gifts& input_data,
@@ -74,11 +75,11 @@ namespace heuristics {
 	void heuristic_chooser
 	(
 		bool heuris_inc, bool heuris_dec, bool heuris_rand, bool heuris_mix,
-		size_t n_threads, size_t rand_times,
-		size_t stop_at, bool stop_time, double stop_when,
+		uint32 n_threads, uint32 rand_times,
+		uint32 stop_at, bool stop_time, double stop_when,
 		bool verbose,
 		const gifts& input_data,
-		wrapped_boxes& solution, size_t& n_sols,
+		wrapped_boxes& solution, uint32& n_sols,
 		bool& error
 	);
